Separate dev_tools, some scripts are not necessary to be released
