+++ dev_tools/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: GPL-3.0-only
