+++ dev_tools/generate_manifest.py
